+++ modules/acid/src/main/ts/ephox/acid/gui/components/HueSlider.ts
@@ -1,5 +1,6 @@
-import { AlloyComponent, AlloyTriggers, Behaviour, Focusing, SketchSpec, Slider } from '@ephox/alloy';
+import { AlloyComponent, AlloyTriggers, Behaviour, Focusing, SketchSpec, Slider, Tabstopping } from '@ephox/alloy';
 import { Fun } from '@ephox/katamari';
+import { Attribute } from '@ephox/sugar';
 
 import { sliderUpdate } from '../ColourEvents';
 
@@ -29,7 +30,11 @@ const sliderFactory = (translate: (key: string) => string, getClass: (key: strin
       tag: 'div',
       classes: [ getClass('hue-slider') ],
       attributes: {
-        role: 'presentation'
+        'role': 'presentation',
+        'aria-role': 'slider',
+        'aria-valuemin': 0,
+        'aria-valuemax': 360,
+        'aria-valuenow': 120,
       }
     },
     rounded: false,
@@ -42,10 +47,12 @@ const sliderFactory = (translate: (key: string) => string, getClass: (key: strin
       thumb
     ],
     sliderBehaviours: Behaviour.derive([
+      Tabstopping.config({ }),
       Focusing.config({ })
     ]),
 
     onChange: (slider: AlloyComponent, _thumb: any, value: any) => {
+      Attribute.set(slider.element, 'aria-valuenow', Math.floor(360 - (value * 3.6)));
       AlloyTriggers.emitWith(slider, sliderUpdate, {
         value
       });