+++ modules/acid/src/main/ts/ephox/acid/gui/components/SaturationBrightnessPalette.ts
@@ -1,5 +1,6 @@
 import { AlloyComponent, AlloyTriggers, Behaviour, Composing, Focusing, Sketcher, SketchSpec, Slider, SliderTypes, UiSketcher } from '@ephox/alloy';
-import { Fun, Optional } from '@ephox/katamari';
+import { Fun, Optional, Type } from '@ephox/katamari';
+import { Attribute } from '@ephox/sugar';
 
 import { Hex } from '../../api/colour/ColourTypes';
 import * as HsvColour from '../../api/colour/HsvColour';
@@ -74,6 +75,7 @@ const paletteFactory = (_translate: (key: string) => string, getClass: (key: str
   const setPaletteThumb = (slider: AlloyComponent, hex: Hex): void => {
     const hsv = HsvColour.fromRgb(RgbaColour.fromHex(hex));
     Slider.setValue(slider, { x: hsv.saturation, y: 100 - hsv.value });
+    Attribute.set(slider.element, 'aria-valuetext', `Saturation ${hsv.saturation}%, Brightness ${hsv.value}%`);
   };
 
   const factory: UiSketcher.SingleSketchFactory<SaturationBrightnessPaletteDetail, SaturationBrightnessPaletteSpec> = (_detail): SketchSpec => {
@@ -83,6 +85,9 @@ const paletteFactory = (_translate: (key: string) => string, getClass: (key: str
     });
 
     const onChange = (slider: AlloyComponent, _thumb: AlloyComponent, value: number | SliderTypes.SliderValue) => {
+      if (!Type.isNumber(value)) {
+        Attribute.set(slider.element, 'aria-valuetext', `Saturation ${Math.floor(value.x)}%, Brightness ${Math.floor(100 - value.y)}%`);
+      }
       AlloyTriggers.emitWith(slider, ColourEvents.paletteUpdate, {
         value
       });
@@ -104,7 +109,8 @@ const paletteFactory = (_translate: (key: string) => string, getClass: (key: str
       dom: {
         tag: 'div',
         attributes: {
-          role: 'presentation'
+          'role': 'presentation',
+          'aria-valuetext': 'Saturation 0%, Brightness 0%'
         },
         classes: [ getClass('sv-palette') ]
       },